uploader/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
ffmpeg-python==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1